requirements.txt
flask>=3.0.0
opencv-python-headless>=4.9.0.80
ultralytics>=8.1.0
tensorflow>=2.16.1
numpy>=1.26.4
mysql-connector-python>=8.3.0
Werkzeug>=3.0.1
python-dotenv>=1.0.0
twilio>=8.0.0
pyngrok>=7.1.0
qrcode>=7.4.2
requests>=2.31.0

